feat: rewrite file in ECMAScript notation
